// src/app/models/article.model.ts
export interface Article {
    id: string;
    title?: string;
    blocks: ArticleBlock[];
    urlNext?: string;
    urlPrevious?: string;
}

export interface ArticleBlock {
    id: string;
    title?: string;
    elements: ArticleElem[];
}

export interface ArticleElem {
    id: string;
    type: ArticleElemType;
    content?: string | string[];
}

export enum ArticleElemTypeEnum {
    text = "TEXT",
    olist = "ORDERED_LIST",
    ulist = "UNORDERED_LIST",
    title = "TITLE",
    link = "LINK"
}

export type ArticleElemType = ArticleElemTypeEnum.text | ArticleElemTypeEnum.olist | ArticleElemTypeEnum.ulist | ArticleElemTypeEnum.title | ArticleElemTypeEnum.link;